[#130389339] always create files for director password and CA cert
Some users may elect to use --print-manifest to generate a manifest
and to use bosh-init to do the deployment.  This way they can keep
the manifest around in case they need to modify the deployment.

This change ensures that the password and cert are written to disk
even when using print manifest, so that users don't have to grep
the manifest for this information.