Remove library for SHA512
github.com/kress/osutil was being used to calculate a SHA512 checksum.
This package was relocated to github.com/tredoe/osutil, and all
commit history was erased, leaving us pointing at a non-existent
commit.

Go's stdlib hash a SHA512 hash so there is no need for a
third-party library.  Remove the reference to this library
and reimplement the functionality using the standard library.